Split CodeEditor component into two components
Currently the component is taking in context state, figuring out which
commands need to be run, executing the commands and highlighting
lines/letters for consistent feedback to user alongside the gesture pad.
This logic and view and should be split up. Create a component which
accepts the command that needs to executed and let the parent component
work out what the command is.