Use command prop to execute commands in code editor
The CodeMirrorEditor doesn't need to know about gestures, this removes a
dependency between interface and code editor. Instead it should use the
command passed down from it's parent component. As request animation is
in the root Interface component, the code editor components are being
rerendered with each increment of count causing the command prop to
passed to the CodeMirrorEditor continuously. Add a check for the command
id, as temp solution until raf is moved. Also remove the id added to
gesture id as a previous temp solution for this issue.