Move request animation frame to interface context
Request animation frame is currently being set in the Interface
component causing all child components to rerender on each tick. To
avoid this and only allow the view components that are using canvas to
subscribe to the count, add it to the context state. Also clean up
unused functions and other code.