Fix scrolling to end of visible code text
The code container is not the same size as the viewable text. This is a
result of the gesture pad being displayed over the text which results in
the cursor not being visible when scrolled to end of the doc. The height
of the code text container - CodeMirror class - should be the reamining
space after the header and GesturePad heights have been set.