Fix bug causing the cursor to be unmarked after an operation is executed
The user needs to know where they are located in the document at all
times. When an operation has been entered, the line is highlighted so
there is still an indication of line position. However, after the
operation has been executed, the state is not reset and therefore the
cursor is not displayed. Change this to check that the user is not
active.