Fix cursor moving when touch grid
The bug is that when the user finishes a gesture and lets go of the or touches down on the gesture pad, the move command is executed. This is temp fix for this bug. We are keeping track of the current gesture and last matched gesture so that we can reset the gesture when the user is inactive.