Move cursor on first single gesture
Currently, the a gesture is not recognised if there is no previous
position to compare it to. This happens when the user is moving from a
inactive state. It prevents small one box incremental movements from
registering which provides misleading feedback to the user. This commit
includes an additional (related) bug fix; the previous position should
only be taken into account by the gesture reader if it's an adjacent
position. This prevents restarts from a different part of the grid
causing immediate cursor movement.