Add a restart button to the header
We want to measure if a user gets stuck and is unable to continue the
task. This can happen if they delete a line they need. Rather than
hitting refresh, they should be encouraged to hit the restart button
which we can measure to see how often users get into this state.