Add header buttons to navigate task state
To improve user experience, all information can be presented in the
simplest way possible - in the code editor. That will require having
multiple states of the task - instructions, code, and a confirmation
that they have completed the task. Buttons in the header will control
these and the code editor will change the text in the editor.