Edit header gesture icons
The icons in the header serve an important purpose in guiding the user. Equally important are the different colours used in the icons and to highlight the text, as this provides feedback for the actions they take in gesture pad. This change updates the colours of the cut and paste icons and renames the yank icon to copt, to be consistent with the copy in the code editor.