Add spaces to empty lines and highlight line on paste
Based on feedback from users, entering an operation gesture is not
clear. The cursor is only displayed on lines with characters (including
spaces) so this will make it visible on all lines. The paste operation
is also not visible. These changes should improve the user experience
when adding an operation.