Fix component rerender bug
The GesturePad component was update gesture to an empty obj when the
user was inactive. To fix this, when the use is not active update the
mode to 'Inactive' and check the mode before updating the gesture state.
Also remove userActive from being passed to the Code component which
doesn't need to be aware of the user's state.