Remove Code component
The purpose behind having a parent Code component was to add a layer
between the gestures and the commands execute by the code editor and
reduce the dependency between the two. This caused performance issues as
the implementation was not clear. Will need to be revisted once tests
are added and I have better confidnece in not breaking existing code.